build: scaffold multi-app platform structure and extract router
Install vite-tsconfig-paths, add @fabkit/* path aliases to tsconfig,
create src/platform/, src/shared/, and src/apps/card-creator/ stubs,
and extract router instance to src/platform/router.tsx to break the
import cycle between bug-report and main.
