+++ .gitignore
@@ -34,4 +34,5 @@ CHANGELOG.md
 # Local AI files
 .claude/settings.local.json
 .claude/worktrees/
+.claude/plans/
 .playwright-mcp/
