+++ package.json
@@ -64,7 +64,8 @@
 		"typescript": "~6.0.3",
 		"vite": "npm:rolldown-vite@7.2.5",
 		"vite-plugin-open-graph": "^2.0.6",
-		"vite-plugin-pwa": "^1.2.0"
+		"vite-plugin-pwa": "^1.2.0",
+		"vite-tsconfig-paths": "^6.1.1"
 	},
 	"overrides": {
 		"vite": "npm:rolldown-vite@7.2.5"
