+++ src/apps/card-creator/index.ts
@@ -1,0 +1,1 @@
+export {};
